requirements.txt
lightning==2.1.2
notebook
numpy==1.26.3
gdown==5.0.0
seaborn
matplotlib
tensorboard==2.15.1
torch==2.1.1
torchmetrics==1.2.1
torchvision==0.16.1
wandb==0.16.1
parameterized==0.9.0
python-dotenv
